Rename UnitaryDynamics to Dynamics
The name only contained unitary for historical reasons, nothing about
the implementation is restricted to unitary dynamics.
