+++ CMakeLists.txt
@@ -56,6 +56,8 @@ endif()
 
 
 set (NETKET_SOURCES ${NETKET_SOURCE_DIR}/netket.cc)
+set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Dynamics)
+set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Dynamics/TimeStepper)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Graph)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Hamiltonian)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Hamiltonian/MatrixWrapper)
@@ -68,8 +70,6 @@ set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Observable)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Sampler)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Optimizer)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Stats)
-set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/UnitaryDynamics)
-set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/UnitaryDynamics/TimeStepper)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Utils)
 
 add_executable(netket ${NETKET_SOURCES})
