+++ NetKet/Dynamics/time_evolution.hpp
